Moved all computation logic to viewmodel

+++ app/src/main/java/ro/bogdanmunteanu/currencyconverter/data/model/BaseCurrencyRow.kt
@@ -1,3 +1,5 @@
 package ro.bogdanmunteanu.currencyconverter.data.model
 
-data class BaseCurrencyRow(override val isoCode:String, override val name: String, override val rate:Double, override val flagUrl:String) : CurrencyRate()
+import java.math.BigDecimal
+
+data class BaseCurrencyRow(override val isoCode:String, override val name: String, override val rate:BigDecimal, override val flagUrl:String) : CurrencyRate()
