+++ app/src/main/java/ro/bogdanmunteanu/currencyconverter/data/model/CurrencyRate.kt
@@ -1,8 +1,10 @@
 package ro.bogdanmunteanu.currencyconverter.data.model
 
+import java.math.BigDecimal
+
 abstract class CurrencyRate {
     abstract val isoCode:String
     abstract val name: String
-    abstract val rate:Double
+    abstract val rate: BigDecimal
     abstract val flagUrl:String
 }
