+++ app/src/main/java/ro/bogdanmunteanu/currencyconverter/data/model/CurrencyRow.kt
@@ -1,3 +1,5 @@
 package ro.bogdanmunteanu.currencyconverter.data.model
 
-data class CurrencyRow(override var isoCode:String, override var name: String, override var rate:Double, override var flagUrl:String) : CurrencyRate()
+import java.math.BigDecimal
+
+data class CurrencyRow(override var isoCode:String, override var name: String, override var rate:BigDecimal, override var flagUrl:String) : CurrencyRate()
